# Likelihood checks speed up.