```
=========================== slowest 5 test durations ===========================
417.60s call     tests/test_likelihoods.py::TestLikelihoodChecks::test_likelihood_checks
166.69s call     tests/test_ekerns.py::TestKernExpQuadrature::test_exKxz_pairwise
148.19s call     tests/test_gplvm.py::TestBayesianGPLVM::test_kernelsActiveDims
65.70s call     tests/test_hmc.py::CheckTrainingVariableState::test_multiple_runs
48.64s call     tests/test_coregion.py::TestEquivalence::test_predicts
```

Likelihood tests take extremely long time. This PR fixes that by using `pytest` parameterize feature and running each likelihood example in different context.