@markvdw 2017-11-25
This looks good, but it moves away from using `unittest` as the code, right? Is this the first part of the code where this happens? Can there be any negative consequences of this?