@awav 2017-11-25
@markvdw, it is even better that we moves away from `unittests`. In this particular case `unittests` provide unnecessary level of abstraction, what's important that we have two distinct tests `pass` and `fail`, and data which we pass - done. Moreover `pytest` is much nicer and provides good features, which we can use, but some of them are not compatible with `unittests`.