@markvdw 2017-11-25
Ok sure, as long as we're happy doing this, and the instructions on how to run the tests in the readme still works.